fix: DRY waffle config
