+++ packages/cf-funding-protocol-contracts/waffle.js
@@ -1,34 +1,3 @@
-const request = require('request');
-
-require("dotenv").config();
-
-var waffleConfig = {
-  "npmPath": "../../node_modules",
-  "legacyOutput": true,
-  "compilerOptions": {
-    "optimizer": {
-      "enabled": true,
-      "runs": 200
-    },
-    "evmVersion": "petersburg",
-    "metadata": {
-      "useLiteralContent": true
-    },
-    "outputSelection": {
-      "*": {
-        "*": [
-          "metadata", "evm.bytecode", "evm.bytecode.sourceMap", "abi"
-        ],
-      },
-    }
-  }
-};
-
-var selectSolc = () => {
-  if (process.env.NATIVE_SOLC == "true") {
-    waffleConfig.compiler = "native";
-  }
-  return waffleConfig;
-}
+const selectSolc = require("../../waffle.js");
 
 module.exports = selectSolc();
